backend/requirements.txt
fastapi
pydantic>=2.0
pydantic-settings>=2.0
httpx>=0.27
python-dotenv>=1.0
uvicorn>=0.30
edge-tts>=6.1
PyPDF2>=3.0
python-multipart>=0.0.6
sentence-transformers>=3.0.1
faiss-cpu>=1.8.0
numpy>=1.26.4
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
